+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Support for delayed stack trace loading
 - Basic support for exception breakpoints
 - Support for conditional breakpoints
+- Support for hit conditional breakpoints
 
 ## 0.1.2